Refactor: 도메인 Member -> User로 변경
