+++ src/main/java/com/example/nzgeneration/domain/errorreport/ErrorReport.java
@@ -2,7 +2,7 @@ package com.example.nzgeneration.domain.errorreport;
 
 import com.example.nzgeneration.global.utils.BaseTimeEntity;
 import com.example.nzgeneration.domain.trashcan.Trashcan;
-import com.example.nzgeneration.domain.member.Member;
+import com.example.nzgeneration.domain.user.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -23,7 +23,7 @@ public class ErrorReport extends BaseTimeEntity {
     private Long id;
 
     @ManyToOne
-    private Member errorReportMember;
+    private User errorReportUser;
 
     private String content;
