--- src/main/java/com/example/nzgeneration/domain/member/MemberRepository.java
@@ -1,8 +1,0 @@
-package com.example.nzgeneration.domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long>
-{
-
-}
