+++ src/main/java/com/example/nzgeneration/domain/user/UserController.java
@@ -1,0 +1,9 @@
+package com.example.nzgeneration.domain.user;
+
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+
+}
