+++ src/main/java/com/example/nzgeneration/domain/user/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.nzgeneration.domain.user;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+
+}
