+++ src/main/java/com/example/nzgeneration/global/security/CurrentUserArgumentResolver.java
@@ -1,7 +1,7 @@
 package com.example.nzgeneration.global.security;
 
-import com.example.nzgeneration.domain.member.Member;
-import com.example.nzgeneration.domain.member.MemberRepository;
+import com.example.nzgeneration.domain.user.User;
+import com.example.nzgeneration.domain.user.UserRepository;
 import com.example.nzgeneration.global.common.response.code.status.ErrorStatus;
 import com.example.nzgeneration.global.common.response.exception.GeneralException;
 import lombok.RequiredArgsConstructor;
@@ -16,12 +16,12 @@ public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolve
 
     private final JwtTokenProvider jwtTokenProvider;
 
-    private final MemberRepository memberRepository;
+    private final UserRepository userRepository;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
         return parameter.getParameterAnnotation(CurrentUser.class) != null
-            && parameter.getParameterType().equals(Member.class);
+            && parameter.getParameterType().equals(User.class);
     }
 
     @Override
